Add/remove modules now add/removes associated folders

+++ php/removeModule.php
@@ -1,17 +1,37 @@
 <?php
 
-include_once("autoload.include.php");
+include_once "autoload.include.php";
 require_once "mypdo.include.php";
 require_once "utils.php";
 
 if(Admin::isConnected()){
 	if($_SERVER["REQUEST_METHOD"] == "POST"){
-		$stmt = myPDO::getInstance()->prepare(<<<SQL
-			DELETE FROM module WHERE idModule = :idMod
+
+		$id = secureInput($_POST['idMod']);
+		$idInt = (int)$id;
+		$module = Module::getModuleById($idInt);
+
+		if($module->getImgMod() !== null){
+			
+			unlink("../" . $module->getImgMod());
+			deleteDirectory("../docs/". $module->getLibModule());
+
+				$stmt = myPDO::getInstance()->prepare(<<<SQL
+				DELETE FROM module WHERE idModule = :idMod
 SQL
 );
+				$stmt->execute(array(':idMod' => $id));
 
-		$stmt->execute(array(':idMod' 	=> secureInput($_POST['idMod'])));
+		}else{
+		
+			$stmt = myPDO::getInstance()->prepare(<<<SQL
+				DELETE FROM module WHERE idModule = :idMod
+SQL
+);
+				$stmt->execute(array(':idMod' => $id));
+		}
+
+		
 
 	}
 }
